+++ packages/utils/package.json
@@ -23,6 +23,6 @@
   "dependencies": {
     "axios": "^1.5.1",
     "semver": "^7.5.4",
-    "url-join": "^5.0.0"
+    "url-join": "^2.0.5"
   }
 }
